fix: move deployment order into contract registry
